feat(mobile): add more sound customization for add sound
